requirements.txt

-e file:///home/runner/work/Crawler/Crawler/backend/lna-db ; python_version >= "3.11" and python_version < "4.0"
annotated-types==0.7.0 ; python_version >= "3.11" and python_version < "4.0"
anyio==4.8.0 ; python_version >= "3.11" and python_version < "4.0"
azure-functions==1.21.3 ; python_version >= "3.11" and python_version < "4.0"
beanie==1.29.0 ; python_version >= "3.11" and python_version < "4.0"
beautifulsoup4==4.13.3 ; python_version >= "3.11" and python_version < "4.0"
certifi==2025.1.31 ; python_version >= "3.11" and python_version < "4.0"
cfgv==3.4.0 ; python_version >= "3.11" and python_version < "4.0"
click==8.1.8 ; python_version >= "3.11" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0" and platform_system == "Windows"
coverage==7.6.1 ; python_version >= "3.11" and python_version < "4.0"
dataclasses-json==0.5.7 ; python_version >= "3.11" and python_version < "4.0"
distlib==0.3.9 ; python_version >= "3.11" and python_version < "4.0"
dnspython==2.7.0 ; python_version >= "3.11" and python_version < "4.0"
email-validator==2.2.0 ; python_version >= "3.11" and python_version < "4.0"
fastapi==0.115.11 ; python_version >= "3.11" and python_version < "4.0"
feedparser==6.0.11 ; python_version >= "3.11" and python_version < "4.0"
filelock==3.17.0 ; python_version >= "3.11" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.11" and python_version < "4.0"
httpcore==1.0.7 ; python_version >= "3.11" and python_version < "4.0"
httpx==0.28.1 ; python_version >= "3.11" and python_version < "4.0"
identify==2.6.9 ; python_version >= "3.11" and python_version < "4.0"
idna==3.10 ; python_version >= "3.11" and python_version < "4.0"
iniconfig==2.0.0 ; python_version >= "3.11" and python_version < "4.0"
langdetect==1.0.9 ; python_version >= "3.11" and python_version < "4.0"
lazy-model==0.2.0 ; python_version >= "3.11" and python_version < "4.0"
libcst==1.6.0 ; python_version >= "3.11" and python_version < "4.0"
marshmallow-enum==1.5.1 ; python_version >= "3.11" and python_version < "4.0"
marshmallow==3.26.1 ; python_version >= "3.11" and python_version < "4.0"
mongomock-motor==0.0.35 ; python_version >= "3.11" and python_version < "4.0"
mongomock==4.3.0 ; python_version >= "3.11" and python_version < "4.0"
motor==3.7.0 ; python_version >= "3.11" and python_version < "4.0"
mypy-extensions==1.0.0 ; python_version >= "3.11" and python_version < "4.0"
nodeenv==1.9.1 ; python_version >= "3.11" and python_version < "4.0"
packaging==24.2 ; python_version >= "3.11" and python_version < "4.0"
platformdirs==4.3.6 ; python_version >= "3.11" and python_version < "4.0"
pluggy==1.5.0 ; python_version >= "3.11" and python_version < "4.0"
pre-commit==4.1.0 ; python_version >= "3.11" and python_version < "4.0"
psutil==7.0.0 ; python_version >= "3.11" and python_version < "4.0"
pydantic-core==2.27.2 ; python_version >= "3.11" and python_version < "4.0"
pydantic==2.10.6 ; python_version >= "3.11" and python_version < "4.0"
pygments==2.19.1 ; python_version >= "3.11" and python_version < "4.0"
pymongo==4.11.2 ; python_version >= "3.11" and python_version < "4.0"
pyre-check==0.9.23 ; python_version >= "3.11" and python_version < "4.0"
pyre-extensions==0.0.32 ; python_version >= "3.11" and python_version < "4.0"
pyright==1.1.396 ; python_version >= "3.11" and python_version < "4.0"
pytest-asyncio==0.24.0 ; python_version >= "3.11" and python_version < "4.0"
pytest-cov==5.0.0 ; python_version >= "3.11" and python_version < "4.0"
pytest==8.3.5 ; python_version >= "3.11" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.11" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.11" and python_version < "4.0"
pytz==2025.1 ; python_version >= "3.11" and python_version < "4.0"
pyyaml==6.0.2 ; python_version >= "3.11" and python_version < "4.0"
ruff==0.9.10 ; python_version >= "3.11" and python_version < "4.0"
sentinels==1.0.0 ; python_version >= "3.11" and python_version < "4.0"
sgmllib3k==1.0.0 ; python_version >= "3.11" and python_version < "4.0"
six==1.17.0 ; python_version >= "3.11" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.11" and python_version < "4.0"
soupsieve==2.6 ; python_version >= "3.11" and python_version < "4.0"
starlette==0.46.1 ; python_version >= "3.11" and python_version < "4.0"
tabulate==0.9.0 ; python_version >= "3.11" and python_version < "4.0"
testslide==2.7.1 ; python_version >= "3.11" and python_version < "4.0"
toml==0.10.2 ; python_version >= "3.11" and python_version < "4.0"
typeguard==2.13.3 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.11" and python_version < "4.0"
typing-inspect==0.9.0 ; python_version >= "3.11" and python_version < "4.0"
uvicorn==0.34.0 ; python_version >= "3.11" and python_version < "4.0"
virtualenv==20.29.3 ; python_version >= "3.11" and python_version < "4.0"
